+++ lib/src/response/forum/exporter.dart
@@ -5,6 +5,6 @@
 export 'package:duolingo4d/src/response/forum/comment.dart';
 export 'package:duolingo4d/src/response/forum/user.dart';
 
-export 'package:duolingo4d/src/response/forum/comments/forum_comments_response.dart';
+export 'package:duolingo4d/src/response/forum/comment/forum_comment_response.dart';
 export 'package:duolingo4d/src/response/forum/topics/forum_topics_response.dart';
 export 'package:duolingo4d/src/response/forum/topic/forum_topic_response.dart';